requirements.txt
wxPython>=4.2.0
requests>=2.28.0
pydub>=0.25.1
python-vlc>=3.0.18122
mutagen>=1.45.1
tqdm>=4.65.0
numpy>=1.24.0
matplotlib>=3.7.0
pyaudio>=0.2.13
qrcode>=7.4.2
Pillow>=9.5.0
